Added Powerup! This initial powerup gives you an extra life if you catch it with your paddle.
NOTE: This is faaaaaaaaaar from good. Currently you can only have 1 powerup at a time, so if the game gives you another powerup it will reset the location of the powerup. 
I use math.random to determine if you get a powerup when you break a block.
The way I determine the initial location of the powerup is shoddy at best.
Needs to be expanded for additional powerups